Cleanup
Just a few followup fixes from the last PR

- Move most of the command handlers code into a separate file to
  further reduce the size of server.go. Not exactly proper
  separation / isolation yet, but it's a first step.
- Remove parser code for command argument format no longer used
